% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateRuntimes.R
\name{aggregateRuntimes}
\alias{aggregateRuntimes}
\title{Runtime aggregation.}
\usage{
aggregateRuntimes(df, group.by = c("prob", "solver"),
  runtime.col = "runtime", scores, ...)
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data frame with columns \dQuote{prob}, \dQuote{solver} and \dQuote{runtime}.}

\item{group.by}{[\code{character(1)}]\cr
Which variables shall be used for grouping? Default is \code{c("prob", "solver")}.
This parameter is useful if the naming of the columns deviates from the defaults.}

\item{runtime.col}{[\code{character(1)}]\cr
Column name of column which contains the running times.
Default is \code{runtime}.
This parameter is useful if the naming of the columns deviates from the defaults.}

\item{scores}{[\code{list(list)}]\cr
Named list of lists. The names on the first layer are passed as argument \code{score}
to function \code{\link{getScore}}. Each element of the list is again an (eventually empty) named list
with the paramters for the corresponding score, e.g., penalty factor \code{f} for PAR-score.}

\item{...}{[any]
Not used.}
}
\value{
[data.frame] Data frame in long format. Basically \code{df} for each conbination of
  score and its parameters.
}
\description{
Given a data frame of pure running times until a certain goal
has been reached, e.g., the optimum was found, computes different performance
scores by problem instance and algorithm/solver.
}
\examples{
\dontrun{
\% setup score parameters
f = c(1, 5, 10)
p = c(0.1, 0.5, 0.9)
cutoff = c(10, 25, 50)
aggr = aggregateRuntimes(df, scores = list(
  "PAR" = list(cutoff = cutoff, f = f),
  "PQR" = list(cutoff = cutoff, p = p, f = f))
)
}
}
