% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colours_clustal}
\alias{colours_clustal}
\title{Clustal colour scheme}
\usage{
colours_clustal(alphabet)
}
\arguments{
\item{alphabet}{a vector of amino acids which specify the order in which the colour palette will
be specified}
}
\description{
Colour scheme described in \url{http://www.ebi.ac.uk/Tools/msa/clustalw2/help/faq.html#24}
}

