% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colours_nt}
\alias{colours_nt}
\title{Nucleotide colour scheme}
\usage{
colours_nt(alphabet)
}
\arguments{
\item{alphabet}{a vector of nucleotides which specify the order in which the colour palette will
be specified}
}
\description{
Colour scheme described in \url{http://www.jalview.org/help/html/colourSchemes/nucleotide.html}
}

