% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colours_zappo}
\alias{colours_zappo}
\title{Zappo colour scheme}
\usage{
colours_zappo(alphabet)
}
\arguments{
\item{alphabet}{a vector of amino acids which specify the order in which the colour palette will
be specified}
}
\description{
Colour scheme described in \url{http://www.jalview.org/version118/documentation.html#zappo}
}

