% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_graphs.R
\name{compare_graphs}
\alias{compare_graphs}
\title{Graph comparison function}
\usage{
compare_graphs(type, extants, ...)
}
\arguments{
\item{type}{operation to perform on the graphs. "union", "intersection" or "complement".}

\item{extants}{filepath to the extant sequence file (clustal or fasta)}

\item{...}{filepaths to graph structures to perform the operation on.}
}
\description{
Takes a list of filepaths to partial order graphs and performs a set operation (i.e. union, intersection or complement)
on the full list of graphs. Saves the resulting partial order graph.\cr
}

