% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{dfError}
\alias{dfError}
\title{Dataframe input error handling}
\usage{
dfError(asrStructure, dataframeName, dataframe, colnames, inference)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}.}

\item{dataframeName}{the name of the dataframe of interest for the parent function}

\item{dataframe}{if provided by the user should contain a value, otherwise NULL}

\item{colnames}{the column names that should be present in the dataframe of interest for the parent function}

\item{inference}{options "Joint" or "Marginal"}
}
\value{
the dataframe required by the parent function or an error
}
\description{
Check inputs provided by user for any not allowed values
}
\examples{

}

