% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{find_positions}
\alias{find_positions}
\title{Identify target positions for mutational analysis}
\usage{
find_positions(asrStructure, max = 20, posFirst = FALSE, distribDF = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}. Set this to NULL
to specify other variables}

\item{max}{the maximum number of positions}

\item{posFirst}{if TRUE, this limits possible combinations at a position to 2}

\item{distribDF}{option to provide the dataframe generated by \code{\link{read_distrib}} rather than use the asrStructure}
}
\value{
a dataframe with three columns representing the position, number of AAs and list of AAs to target for mutation
posDF$Position\cr
posDF$Mode\cr
posDF$AAs\cr
#Example\cr
posDF[1, ] = "186", "3", "H,C,N" - Position 186 is optimised at 3 positions and the top letters are H, C and N\cr
}
\description{
Identify target positions for mutational analysis
}
\examples{
data(asrStructure)

find_positions(asrStructure)
find_positions(asrStructure, max = 50)
find_positions(asrStructure, max = 10, posFirst = TRUE)

#if you want to use a specific set of data
#retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"

distribDF <- read_distrib(NULL, distrib_file = id_distribution)
find_positions(NULL, distribDF = distribDF) #if you want to specify your own file

}

