% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromatograms.R
\name{get_chromatograms}
\alias{get_chromatograms}
\title{Get Chromatogram}
\usage{
get_chromatograms(file)
}
\arguments{
\item{file}{a valid filepath for a \code{.ASR} file}
}
\value{
a list of chromatogram blocks
}
\description{
Extract individual chromatograms for individual channels
}
