% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_extant_sequence}
\alias{get_extant_sequence}
\title{Get a list of extant sequence names}
\usage{
get_extant_sequence(asrStructure)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}
}
\value{

}
\description{
Takes the stored Newick string of the relabelled tree and extracts the extant sequences to be used
in functions like \code{\link{plot_aln}}
}
\examples{

}

