% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_internal_sequence}
\alias{get_internal_sequence}
\title{Get a vector of internal sequence names}
\usage{
get_internal_sequence(asrStructure)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}. Set this to NULL
to specify other variables}
}
\value{

}
\description{
Takes the stored Newick string of the relabelled tree and extracts the internal sequences to be used
in functions like \code{\link{plot_aln}}
}
\examples{

}

