% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_table.R
\name{get_peak_table}
\alias{get_peak_table}
\title{Get Peak Tables}
\usage{
get_peak_table(file)
}
\arguments{
\item{file}{a valid filepath for a \code{.ASR} file}
}
\value{
a list of peak tables
}
\description{
Extract peaks tables from processed chromatograms for individual channels
}
