% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_info.R
\name{get_sample_info}
\alias{get_sample_info}
\title{Get Sample Info}
\usage{
get_sample_info(file)
}
\arguments{
\item{file}{a valid filepath for a \code{.ASR} file}
}
\value{
a \code{data.frame} of sample information
}
\description{
Extract a minimal amount of sample information
}
