% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln.R
\name{get_seq_df}
\alias{get_seq_df}
\title{Format sequence from fasta}
\usage{
get_seq_df(asrStructure, fastaDF = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{fastaDF}{a dataframe generated by \code{\link{read_fasta}}}
}
\value{
a dataframe with three columns representing the amino acid sequences. \cr
Formatted to be passed to \code{\link{plot_aln}}.\cr
seqDF$Label - label/id stored in fasta file\cr
seqDF$Column - number representing the column in the alignment this result belongs to \cr
seqDF$AA - amino acid located in this column belonging to labelled sequence\cr
#example\cr
seqDF[1,] = "N1", "5", "A"  - sequence N1_ contains an A at position 5 in its alignment sequence\cr
}
\description{
Takes the dataframe generated by \code{\link{read_fasta}} and formats it 
so it can be passed to \code{\link{plot_aln}} to be plotted using ggplot2
}
\examples{
data(asrStructure)

get_seq_df(asrStructure)

#OR

#retrieve example file stored in the package
fasta_file <- system.file("extdata", "runASR_aln_full.fa", package="ASR")
#alternatively, specify the filename as a string
#fasta_file <- "id_aln_full.fa"

fastaDF <- read_fasta(NULL, aln_file = fasta_file) # read in the fasta file returned by ASR.jar
get_seq_df(NULL, fastaDF=fastaDF) #fastaDF generated by read_fasta

}

