% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree.R
\name{get_subtree_sequences}
\alias{get_subtree_sequences}
\title{Generate a vector of node labels given the subtree}
\usage{
get_subtree_sequences(asrStructure, node, fastaDF = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{node}{label of node to form root of subtree}

\item{fastaDF}{a dataframe generated by \code{\link{read_fasta}} - only specify this if you want to use a specific dataframe}
}
\value{
a vector containing all nodes/sequences in the subtree
}
\description{
Takes a node and extracts all descendants of node then returns descendants as a list that could, for example, be passed to \code{\link{plot_aln}}
to generate an alignment of the subtree
}
\examples{
data(asrStructure)

get_subtree_sequences(asrStructure, "N1")

#if you want to use a specific set of data
#retrieve example file stored in the package
fasta_file <- system.file("extdata", "runASR_aln_full.fa", package="ASR")
#alternatively, specify the filename as a string
#fasta_file <- "id_aln_full.fa"

fastaDF <- read_fasta(NULL, aln_file = fasta_file)
get_subtree_sequences(NULL, "N1", fastaDF = fastaDF) # if you want to use a specific dataframe

}

