% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loadASR}
\alias{loadASR}
\title{Load existing \code{\link{runASR}} or ASR.jar output}
\usage{
loadASR(id, tree_file = NULL, aln_file = NULL, distrib_file = NULL,
  plot = TRUE)
}
\arguments{
\item{id}{the identifier used when running ASR.jar - if files exist in a different directory use: "dir/id" or "dir1/dir2/id" as input. 
If you want to use a specific combination of files, put id as NULL}

\item{tree_file}{a newick string representation of the re-labelled tree generated by ASR.jar [OPTIONAL]}

\item{aln_file}{a fasta formatted multiple sequence alignment generated by ASR.jar [OPTIONAL]}

\item{distrib_file}{a matrix generated by ASR.jar that represents distributions across amino acids at columns throughout the alignment [OPTIONAL]}

\item{plot}{specifies whether or not to save the plots generated when calling this function}
}
\value{
asrStructure - a named list containing all data structures and files required to run all other functions in package. If plot = TRUE, 
figures associated with generated output (see description) will be generated.
All functions can take asrStructure as input (first argument), to use specific objects, rather than those stored in asrStructure, 
set the first argument to NULL.\cr
asrStructure:\cr
$fileNames - string representations of file names (nested list)\cr
 $fileNames$FastaFile\cr
 $fileNames$Distrib\cr
 $fileNames$Tree\cr
$loadedFiles - loaded/read in files from FileNames \cr
 $loadedFiles$tree - loaded tree file\cr
 $loadedFiles$alignment - loaded alignment/fasta file\cr
 $loadedFiles$distribution - loaded distribution file \cr
$fastaDF - processed fasta file (see \code{\link{read_fasta}})\cr
$seqDF - reformatted output from \code{\link{read_fasta}} (see \code{\link{get_seq_df}})\cr
$distribProb - processed distribution file (see \code{\link{read_distrib}})\cr
$distribHeight - processed probability distribution (see \code{\link{logo_height}})\cr
}
\description{
A wrapper to initialise required data structure to process and visualise data. Run this code if you have previously
generated output from runASR or ASR.jar and want to load the required structures.
}
\examples{
loadASR("test_joint") # if you have run runASR with id = "test_joint" and want to reload the output
loadASR("test_joint", plot = FALSE) #Don't save any figure output
#if you want to specify files to use
loadASR(NULL, tree_file = "id_new_tree.txt", aln_file = "id_aln_full.fa") 
#if you want to combine results from multiple runs of runASR
loadASR(NULL, tree_file = "id_new_tree.txt", aln_file = "id_aln_full.fa",
distrib_file = "id2_distribution.txt") 

}

