% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_aln.R
\name{logo_height_aln}
\alias{logo_height_aln}
\title{Calculate the height of letters in columns}
\usage{
logo_height_aln(asrStructure, seqDF = NULL, alphabet = NULL,
  alphabetType = NULL, sequences = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}. Set this to NULL
to specify other variables}

\item{seqDF}{the dataframe generated by \code{\link{get_seq_df}}}

\item{alphabet}{choice of AA, DNA or RNA}

\item{sequences}{the sequences to be used when calculating column sequence heights}
}
\value{
a dataframe with three columns representing the height of each AA at each column in the alignment. Formatted to
be passed to \code{\link{plot_logo_aln}}\cr
heightDF$AA\cr
heightDF$Column \cr
heightDF$Height \cr
#Example\cr
heightDF[1, ] = "A", "5", "1.17" - An A in column 5 has a height of 1.17\cr
}
\description{
Takes the seqDF (see \code{\link{get_seq_df}}) data frame and calculates first relative frequence
and then logo height for each amino acid in each column
}
\examples{
data(asrStructure)

logo_height_aln(asrStructure)

alphabet <- c("H", "C", "N", "P", "R") #Order is irrelevant
logo_height_aln(asrStructure, alphabet=alphabet)

#this method attempts to detect the alphabet but allows the user to specify
logo_height_aln(asrStructure, alphabet="DNA")

#retrieve example file stored in the package
fasta_file <- system.file("extdata", "runASR_aln_full.fa", package="ASR")
#alternatively, specify the filename as a string
#fasta_file <- "id_aln_full.fa"

# read in the fasta file of choice
fasta <- read_fasta(NULL, aln_file = fasta_file) 
# convert this to a dataframe that can be read by plot_aln
seqDF <- get_seq_df(NULL, fastaDF = fasta) 
logo_height_aln(NULL, seqDF = seqDF) #if you want to specify your own file

}

