% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_distrib.R
\name{logo_height_distrib}
\alias{logo_height_distrib}
\title{Calculate the height of letters in columns}
\usage{
logo_height_distrib(asrStructure, distribDF = NULL, alphabet = NULL,
  alphabetType = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}. Set this to NULL
to specify other variables}

\item{distribDF}{the dataframe generated by \code{\link{read_distrib}}}

\item{alphabet}{an alphabet which represents the lettering of the distribution of interest}

\item{alphabetType}{choice of AA, DNA, RNA or User}
}
\value{
a dataframe with three columns representing the height of each AA at each column in the alignment. Formatted to
be passed to \code{\link{plot_logo_distrib}}\cr
heightDF$AA\cr
heightDF$Column \cr
heightDF$Height \cr
#Example\cr
heightDF[1, ] = "A", "5", "1.17" - An A in column 5 has a height of 1.17\cr
}
\description{
Takes the distribution data frame and calculates a logo height for each amino acid in each column
}
\examples{
data(asrStructure)

logo_height_distrib(asrStructure)

alphabet <- c("H", "C", "N", "P", "R") #Order is irrelevant
logo_height_distrib(asrStructure, alphabet=alphabet)

#this method attempts to detect the alphabet but allows the user to specify
logo_height_distrib(asrStructure, alphabet="DNA")

#if you want to use a specific set of data
#retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"

distribDF <- read_distrib(NULL, distrib_file = id_distribution)
logo_height_distrib(NULL, distribDF = distribDF) #if you want to specify your own file

}

