% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{performance_measures}
\alias{performance_measures}
\alias{getScore}
\alias{getPARScore}
\alias{getPQRScore}
\alias{getPERTScore}
\alias{getSuccessRate}
\alias{getFailureRate}
\alias{getRuntimeQuantile}
\alias{getHVScore}
\title{Performance measures.}
\usage{
getScore(times, score, ...)

getPARScore(times, cutoff = 3600, f = 10, ...)

getPQRScore(times, cutoff = 3600, f = 10, p = 0.5, ...)

getPERTScore(times, cutoff = 3600, f = 1, ...)

getSuccessRate(times, cutoff = 3600)

getFailureRate(times, cutoff = 3600)

getRuntimeQuantile(times, cutoff = 3600, p = 0.5, ...)

getHVScore(times, cutoff = 3600, p = 0.5, normalize = FALSE, ...)
}
\arguments{
\item{times}{[\code{numeric}]\cr
Vector of running times of algorithm runs.}

\item{score}{[\code{character(1)}]\cr
Name of score.}

\item{...}{[any]\cr
Arguments passed down to score functions.}

\item{cutoff}{[\code{numeric(1)}]\cr
Cutoff time for algorithm runs in seconds. Runs with running times exceeding
the cutoff time are categorized as failed. Default is 1 hour.}

\item{f}{[\code{numeric(1)}]\cr
Penalty factor. Each failed run is penalized by \code{cutoff} times \code{f}.
Default is 10, which is quite common in the literature on algorithm selection.}

\item{p}{[\code{numeric(1)}]\cr
Quantile. Default is 0.5.}
}
\value{
[\code{numeric(1)}] Scalar performance value.
}
\description{
Performance measures which are frequently used in the field
of algorithm selection in both the continuous and discrete domains.
}
