% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln.R
\name{plot_aln}
\alias{plot_aln}
\title{Alignment plotting function}
\usage{
plot_aln(asrStructure, seqDF = NULL, type = "colouredText",
  colour = "clustal", columns = NULL, sequences = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}. Set this to NULL
to specify other variables}

\item{seqDF}{a data frame generated by \code{\link{get_seq_df}}}

\item{type}{how the alignment should be displayed, default colouredText,
options: "colour", "text", "both", "colouredText", "logo"
#'NOTE: "logo" uses \code{\link{plot_logo_aln}}}

\item{colour}{the colour scheme selected, default clustal, options: "clustal", "zappo", "taylor", "mixed"}

\item{columns}{a vector containing the column numbers of interest. By default = NULL and all columns are displayed}

\item{sequences}{a vector containing the sequences (by Label) of interest. By default = NULL and all sequences are displayed}
}
\value{
plots a matrix of amino acid sequence given columns and sequences of interest
}
\description{
Creates a plot of the alignment produced by ASR.jar which can be represented as 
tiles, plain text, coloured text or a combination. Colours are specified in \code{colours.R} 
and currently include Clustal and Zappo colour schemes
}
\examples{
data(asrStructure)

plot_aln(asrStructure) # plot using all defaults
#plot alignment displaying plain text
plot_aln(asrStructure, type = "text")
#plot alignment displaying coloured text
plot_aln(asrStructure, type = "colouredText")
#plot alignment displaying coloured boxes
plot_aln(asrStructure, type = "colour")
#plot alignment displaying coloured boxes with overlayed text
plot_aln(asrStructure, type = "both")
#'#plot alignment displaying a boxplot logo
plot_aln(asrStructure, type = "logo")
# plot alignment using zappo colouring (assuming the type argument uses colour) 
plot_aln(asrStructure, colour = "zappo")
# plot alignment using taylor colouring and coloured boxes
#(assuming the type argument uses colour)
plot_aln(asrStructure, colour = "taylor", type = "colour")

#Loading and using raw data files from runASR

#retrieve example file stored in the package
fasta_file <- system.file("extdata", "runASR_aln_full.fa", package="ASR")
#alternatively, specify the filename as a string
#fasta_file <- "id_aln_full.fa"

# read in the fasta file of choice
fasta <- read_fasta(NULL, aln_file = fasta_file) 
# convert this to a dataframe that can be read by plot_aln
seqDF <- get_seq_df(NULL, fastaDF = fasta) 
plot_aln(NULL, seqDF = seqDF) # plot using all defaults

#Plotting specific columns and/or sequences
columns = c(3,4,5,6)
sequences = c("N1", "N3", "N2", "extant_1")
plot_aln(asrStructure, columns = columns)
plot_aln(asrStructure, sequence = sequences)
plot_aln(asrStructure, columns = columns, sequence = sequences)


}

