% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_distrib}
\alias{plot_distrib}
\title{Plot marginal distrib}
\usage{
plot_distrib(asrStructure, distribDF = NULL, type = "colouredText",
  colour = NULL, columns = NULL, aas = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{distribDF}{a dataframe created by \code{\link{read_distrib}}}

\item{type}{how the distribution should be displayed, default plain text.}

\item{colour}{only available in conjunction with type="logo" to specify the colour scheme for AAs. options: "clustal", "zappo", "taylor"
options: "colour", "text", "logo", "colouredText"
NOTE: "logo" uses \code{\link{plot_logo_distrib}}}

\item{columns}{a vector containing the column numbers of interest. By default = NULL and all columns are displayed}

\item{aas}{a vector containing the amino acids of interest. By default = NULL and all amino acids are displayed}
}
\value{
plots a matrix of amino acid probability at given columns in alignment
}
\description{
Displays a heatmap showing the probabilities of each amino acid across all columns in the alignment
}
\examples{
data(asrStructure)

plot_distrib(asrStructure) #if you want to use the information stored in the structure

#To change the display of the plot
plot_distrib(asrStructure, type = "colour")
plot_distrib(asrStructure, type = "text")
plot_distrib(asrStructure, type = "logo")
plot_distrib(asrStructure, type = "colouredText")


#if you want to use a specific set of data
#' #retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"

distribDF <- read_distrib(NULL, id_distribution)
plot_distrib(NULL, distribDF = distribDF)

#to specify particular columns and AAs to include in figure
columns = c(4,5,6,7)
aas = c("G", "P")

plot_distrib(asrStructure, columns = columns)
plot_distrib(asrStructure, aas = aas)
plot_distrib(asrStructure, columns = columns, aas = aas)
#OR
plot_distrib(NULL, distribDF = distribDF, columns = columns)

}

