% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_aln.R
\name{plot_logo_aln}
\alias{plot_logo_aln}
\title{Plot amino acid heights as a barchart to represent logo}
\usage{
plot_logo_aln(asrStructure, heightDF = NULL, colour = "taylor",
  columns = NULL, sequences = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}. Set this to NULL
to specify other variables}

\item{heightDF}{a data frame containing heights generated by \code{\link{logo_height_aln}}}

\item{colour}{the colour scheme to be used to colour AAs}

\item{columns}{a vector containing the column numbers of interest. By default = NULL and all columns are displayed}

\item{sequences}{the sequences to be used when calculating column sequence heights}
}
\value{
plots a variation on a sequence logo based on seqHeights of AAs in columns
}
\description{
Takes the sequence height data frame and plot a barchart for each column using the calculated heights
}
\examples{
data(asrStructure)

plot_logo_aln(asrStructure) #if you want to use the information stored in the structure

#retrieve example file stored in the package
fasta_file <- system.file("extdata", "runASR_aln_full.fa", package="ASR")
#alternatively, specify the filename as a string
#fasta_file <- "id_aln_full.fa"

# read in the fasta file of choice
fasta <- read_fasta(NULL, aln_file = fasta_file) 
# convert this to a dataframe that can be read by plot_aln
seqDF <- get_seq_df(NULL, fastaDF = fasta) 
heightDF <- logo_height_aln(NULL, seqDF = seqDF)
plot_logo_aln(NULL, heightDF = heightDF)

#to specify particular columns to include in figure
columns = c(3,4,5,6)

plot_logo_aln(asrStructure, columns = columns)
#OR
plot_logo_aln(NULL, heightDF = heightDF, columns = columns)

#to change the colouring of AAs
plot_logo_aln(asrStructure, colour = "clustal")
plot_logo_aln(asrStructure, colour = "zappo")

}

