% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_distrib.R
\name{plot_logo_distrib}
\alias{plot_logo_distrib}
\title{Plot amino acid heights as a barchart to represent logo}
\usage{
plot_logo_distrib(asrStructure, heightDF = NULL, colour = "taylor",
  columns = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}}, \code{\link{loadASR}} or \code{\link{reduce_alphabet}}. Set this to NULL
to specify other variables}

\item{heightDF}{a data frame containing heights generated by \code{\link{logo_height_distrib}}}

\item{colour}{the colour scheme to be used to colour AAs}

\item{columns}{a vector containing the column numbers of interest. By default = NULL and all columns are displayed}
}
\value{
plots a variation on a sequence logo based on distribDF of AAs in columns
}
\description{
Takes the distribution height data frame and plot a barchart for each column using the calculated heights
}
\examples{
data(asrStructure)

plot_logo_distrib(asrStructure) #if you want to use the information stored in the structure

#if you want to use a specific set of data
#retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"

distribDF <- read_distrib(NULL, distrib_file = id_distribution)
heightDF <- logo_height_distrib(NULL, distribDF = distribDF)
plot_logo_distrib(NULL, heightDF = heightDF)

#to specify particular columns to include in figure
columns = c(3,4,5,6)

plot_logo_distrib(asrStructure, columns = columns)
#OR
plot_logo_distrib(NULL, heightDF = heightDF, columns = columns)

#to change the colouring of AAs
plot_logo_distrib(asrStructure, colour = "clustal")
plot_logo_distrib(asrStructure, colour = "zappo")
plot_logo_distrib(asrStructure, colour = "mixed")

}

