% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Tree plotting function}
\usage{
plot_tree(asrStructure, tree_file = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{tree_file}{output file from ASR.jar containing a Newick string}
}
\value{
plots a figure containing the tree of interest
}
\description{
Creates a plot of a tree stored as a Newick string
}
\examples{
data(asrStructure)

plot_tree(asrStructure)

#retrieve example file stored in the package
new_tree <- system.file("extdata", "runASR_new_tree.txt", package="ASR")
#alternatively, specify the filename as a string
#new_tree <- "id_new_tree.txt"

plot_tree(NULL, tree_file = new_tree) # if you want to use a specific tree file

}

