% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{plot_tree_seq}
\alias{plot_tree_seq}
\title{Tree plotting function}
\usage{
plot_tree_seq(asrStructure, column, type = "text", colour = "black")
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{column}{the column that you wish to see sequences for}

\item{type}{the way the sequence should be displayed. Options: "text", "shape", "both"}

\item{colour}{the colour scheme selected, default "black", options: "clustal", "zappo", "taylor". Also an option to provide
a user specified colour e.g. colour="red" or colour="#FF0000"}
}
\description{
Creates a plot of a tree stored as a Newick string and labels extant and internal nodes according to their sequence in 
a specified column. AAs can also be coloured using available schemes
}
\examples{
plot_tree_seq(asrStructure, 3)
plot_tree_seq(asrStructure, 3, type="both", colour="clustal")
plot_tree_seq(asrStructure, 3, type="shape", colour="clustal")


}

