% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{plot_tree_seq_label}
\alias{plot_tree_seq_label}
\title{Tree plotting function}
\usage{
plot_tree_seq_label(asrStructure, column, colour = "clustal")
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{tree_file}{output file from ASR.jar containing a Newick string}
}
\value{

}
\description{
Creates a plot of a tree stored as a Newick string
}
\examples{

}

