% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterPerformance.R
\name{plotScatterPerformance}
\alias{plotScatterPerformance}
\title{Performance scatter-plot}
\usage{
plotScatterPerformance(x, value.var = "PAR10", algo.a, algo.b,
  color = "group", shape = "group")
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Data frame with columns at least \dQuote{solver} and \code{value.var}.}

\item{value.var}{[\code{character(1)}]\cr
Name of column which contains the performance values.}

\item{algo.a}{[\code{character(1)}]\cr
Name of first algorithm (x-axis).}

\item{algo.b}{[\code{character(1)}]\cr
Name of second algorithm (y-axis).}

\item{color}{[\code{character(1)}]\cr
Name of column which determines coloring of points, e.g. instance group.}

\item{shape}{[\code{character(1)}]\cr
Name of column which determines the shape of points, e.g. instance group.}
}
\value{
[\code{ggplot}] ggplot2 object.
}
\description{
Generates a scatter-plot of pairwise performances of two solvers.
The function expects a data frame \code{x} which contains performance values
row-wise. The mandatory column \dQuote{solver} encodes the name of the
corresponding algorithm.

The procedure works as follows: the data frame \code{x} column is filtered
to all rows where \code{solver$x} is either \code{algo.a} or \code{algo.b}.
Next, a transformation of the data frame is conducted via \code{\link[reshape2]{dcast}}
to produce a wide data frame with each one column for \code{algo.a} and \code{algo.b}
respectively. These columns contain the performance values indicated by argument
\code{value.var}.
}
