% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{read_distrib}
\alias{read_distrib}
\title{Read in a marginal distribution}
\usage{
read_distrib(asrStructure, distrib_file = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{distrib_file}{distribution file generated by ASR.jar}
}
\value{
a dataframe with three columns representing the distribution at each column in the alignment. Formatted to
be passed to \code{\link{plot_distrib}}.\cr
distribDF$AA\cr
distribDF$Column \cr
distribDF$Probability \cr
#Example\cr
distribDF[1, ] = "A", "5", "0.9" - An A in column 5 has a probability of 0.9\cr
}
\description{
Create a dataframe from the distribution file generated by ASR.jar using marginal inference
}
\examples{
data(asrStructure)

read_distrib(asrStructure)
#retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"
read_distrib(NULL, distrib_file = id_distribution) #if you want to specify your own file

}

