% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read in a fasta alignment file}
\usage{
read_fasta(asrStructure, aln_file = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{aln_file}{aln file generated by ASR.jar}
}
\value{
a dataframe with three columns representing the sequences stored in the fasta file. Formatted to
be passed to \code{\link{get_seq_df}}.\cr
fastaDF$Label - label/id of sequence\cr
fastaDF$Newick - newick string with current node/sequence as root. NULL if extant\cr
fastaDF$Sequence - a single string representing the sequence\cr
#Example\cr
fastaDF[1,] = "N4", "(extant_3:0.05,extant_4:0.05)N4:0.05;", "G-PPGPKT"\cr
}
\description{
Takes a fasta alignment file generate by ASR.jar and formats it into 
a dataframe that can be read by other functions (e.g. \code{\link{get_seq_df}})
}
\examples{
data(asrStructure)

read_fasta(asrStructure)

#if you want to use a specific set of data 
#retrieve example file stored in the package
fasta_file <- system.file("extdata", "runASR_aln_full.fa", package="ASR")
#alternatively, specify the filename as a string
#fasta_file <- "id_aln_full.fa"

read_fasta(NULL, aln_file = fasta_file) # if you want to use a specific file

}

