% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runASR}
\alias{runASR}
\title{Run ASR.jar}
\usage{
runASR(tree_file, aln_file, inf = "Joint", node = NULL, id = "runASR",
  plot = TRUE)
}
\arguments{
\item{tree_file}{a newick string representation of the tree of interest}

\item{aln_file}{a clustal or fasta formatted multiple sequence alignment}

\item{inf}{the inference approach to be selected: "Joint" or "Marginal"}

\item{node}{the node to be queried in inf = "Marginal}

\item{id}{an identifier for the output files}

\item{plot}{specifies whether or not to save the plots generated when calling this function}
}
\value{
asrStructure - a named list containing all data structures and files required to run all other functions in package.\cr
**output will vary depending on input (see description)**\cr
If plot = TRUE, figures associated with generated output (see description) will be generated.\cr
All functions can take asrStructure as input (first argument), to use specific objects, 
rather than those stored in asrStructure, set the first argument to NULL.\cr
asrStructure:\cr
$fileNames - string representations of file names (nested list)\cr
 $fileNames$FastaFile\cr
 $fileNames$Distrib\cr
 $fileNames$Tree\cr
$loadedFiles - loaded/read in files from FileNames \cr
 $loadedFiles$tree - loaded tree file\cr
 $loadedFiles$alignment - loaded alignment/fasta file\cr
 $loadedFiles$distribution - loaded distribution file \cr
$fastaDF - processed fasta file (see \code{\link{read_fasta}})\cr
$seqDF - reformatted output from \code{\link{read_fasta}} (see \code{\link{get_seq_df}})\cr
$seqHeights - processed probability distribution (see \code{\link{logo_height_aln}})\cr
$distribProb - processed distribution file (see \code{\link{read_distrib}})\cr
$distribHeights - processed probability distribution (see \code{\link{logo_height_distrib}})\cr
}
\description{
A wrapper to launch ASR.jar generating all appropriate files according to input. \cr
GENERATE SEQUENCE - requires use of Joint inference - returns a tree and alignment file \cr
GENERATE DISTRIBUTION - requires use of Marginal inference - returns a tree and distribution file \cr 
These two tasks MUST be run SEPARATELY \cr
}
\examples{
#retrieve example file stored in the package
tree <- system.file("extdata", "test_tree.txt", package="ASR")
aln <- system.file("extdata", "test_aln.txt", package="ASR")
#alternatively, specify the filename as a string
#tree <- "tree.txt"
#aln <- "aln.txt"

runASR(tree, aln)
runASR(tree, aln, id = "newID")
runASR(tree, aln, inf = "Joint", id = "testJoint")
runASR(tree, aln, inf = "Marginal", id = "testMarginal")
runASR(tree, aln, inf = "Marginal", node = "N1", id = "testMarginal_N1")

}

