% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln.R
\name{save_aln}
\alias{save_aln}
\title{Alignment plotting and saving function}
\usage{
save_aln(asrStructure, seqDF = NULL, type = "colouredText",
  colour = "clustal", columns = NULL, sequences = NULL, format = "pdf",
  name = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{seqDF}{a data frame generated by \code{\link{get_seq_df}}}

\item{type}{how the alignment should be displayed, default colouredText,
options: "colour", "text", "both", "colouredText", "logo"}

\item{colour}{the colour scheme selected, default clustal, options: "clustal", "zappo", "taylor", "mixed"}

\item{columns}{a vector containing the column numbers of interest. By default = NULL and all columns are displayed}

\item{sequences}{a vector containing the sequences (by Label) of interest. By default = NULL and all sequences are displayed}

\item{format}{specifies what format the figure should be saved in. Options: "pdf" or "png"}

\item{name}{what name would you like the figure saved under?}
}
\description{
Creates a plot of the alignment produced by ASR.jar which can be represented as 
tiles, plain text, coloured text or a combination. Colours are specified in \code{colours.R} 
and currently include Clustal and Zappo colour schemes. This plot is then saved.
}
\examples{
data(asrStructure)

save_aln(asrStructure) # plot using all defaults
#plot alignment displaying plain text
save_aln(asrStructure, type = "colour") 
#'#'#plot alignment displaying a boxplot logo
save_aln(asrStructure, type = "logo")
# plot alignment using zappo colouring (assuming the type argument uses colour)
save_aln(asrStructure, colour = "zappo")

#Loading and using raw data files from runASR

#retrieve example file stored in the package
fasta_file <- system.file("extdata", "runASR_aln_full.fa", package="ASR")
#alternatively, specify the filename as a string
#fasta_file <- "id_aln_full.fa"

# read in the fasta file returned by ASR.jar
fasta <- read_fasta(NULL, aln_file = fasta_file) 
# convert this to a dataframe that can be read by plot_aln
seqDF <- get_seq_df(NULL, fastaDF = fasta)
save_aln(NULL, seqDF = seqDF) # plot using all defaults
#plot alignment displaying plain text
save_aln(NULL, seqDF = seqDF, type = "colour")
# plot alignment using zappo colouring (assuming the type argument uses colour) 
save_aln(NULL, seqDF = seqDF, colour = "zappo") 

#Plotting specific columns and/or sequences
columns = c(3,4,5,6)
sequences = c("N1", "N3", "N2", "extant_1")
save_aln(asrStructure, columns = columns)
save_aln(asrStructure, sequence = sequences)
save_aln(asrStructure, columns = columns, sequence = sequences)

save_aln(asrStructure, format = "png", name = "new_name")

}

