% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{save_distrib}
\alias{save_distrib}
\title{Save marginal distrib plot}
\usage{
save_distrib(asrStructure, distribDF = NULL, type = "colouredText",
  columns = NULL, aas = NULL, format = "pdf", name = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{distribDF}{a dataframe created by \code{\link{read_distrib}}}

\item{type}{how the distribution should be displayed, default plain text.
options: "colour", "text", "logo", "colouredText"
NOTE: "logo" uses \code{\link{logo_height}} and \code{\link{save_logo}}}

\item{columns}{a vector containing the column numbers of interest. By default = NULL and all columns are displayed}

\item{aas}{a vector containing the amino acids of interest. By default = NULL and all amino acids are displayed}

\item{format}{specifies what format the figure should be saved in. Options: "pdf" or "png"}

\item{name}{what name would you like the figure saved under?}
}
\description{
Saves a heatmap showing the probabilities of each amino acid across all columns in the alignment
}
\examples{
data(asrStructure)

save_distrib(asrStructure) #if you want to use the information stored in the structure

#if you want to use a specific set of data
#' #retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"

distribDF <- read_distrib(NULL, id_distribution)
save_distrib(NULL, distribDF = distribDF)

#to specify particular columns and AAs to include in figure
columns = c(4,5,6,7)
aas = c("G", "P")

save_distrib(asrStructure, columns = columns)
save_distrib(asrStructure, aas = aas)
save_distrib(asrStructure, columns = columns, aas = aas)
#OR
save_distrib(NULL, distribDF = distribDF, columns = columns)

save_distrib(asrStructure, format = "png", name = "new_name")

}

