% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{save_logo_distrib}
\alias{save_logo_distrib}
\title{Save amino acid heights as a barchart to represent logo}
\usage{
save_logo_distrib(asrStructure, heightDF = NULL, colour = "taylor",
  columns = NULL, format = "pdf", name = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{heightDF}{a data frame containing heights generated by \code{\link{logo_height}}}

\item{colour}{the colour scheme to be used to colour AAs}

\item{columns}{a vector containing the column numbers of interest. By default = NULL and all columns are displayed}

\item{format}{specifies what format the figure should be saved in. Options: "pdf" or "png"}

\item{name}{what name would you like the figure saved under?}
}
\value{
plots a variation on a sequence logo based on probabilities of AAs in columns
}
\description{
Takes the distribution height data frame and plots then saves a barchart for each column using the calculated heights
}
\examples{
data(asrStructure)

save_logo(asrStructure) #if you want to use the information stored in the structure

#if you want to use a specific set of data
#retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"

distribDF <- read_distrib(NULL, distrib_file = id_distribution)

heightDF <- logo_height(NULL, distribDF = distribDF)
save_logo(NULL, heightDF = heightDF)

#to specify particular columns to include in figure
columns = c(3,4,5,6)

save_logo(asrStructure, columns = columns)
#OR
save_logo(NULL, heightDF = heightDF, columns = columns)

#to change the colouring of AAs
save_logo(asrStructure, colour = "clustal")

save_logo(asrStructure, format = "png", name = "new_name")

}

