% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{save_tree}
\alias{save_tree}
\title{Tree saving function}
\usage{
save_tree(asrStructure, tree_file = NULL, format = "pdf", name = NULL)
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{tree_file}{output file from ASR>jar containing a Newick string}

\item{format}{specifies format to save figure in. Options: "pdf" or "png"}

\item{name}{what name would you like the figure saved under?}
}
\description{
Creates a plot of a tree stored as a Newick string and saves it as an
image using the name of the tree_file argument
}
\examples{
data(asrStructure)

save_tree(asrStructure)
save_tree(asrStructure, format="png")

# if you want to use a specific tree file
#retrieve example file stored in the package
new_tree <- system.file("extdata", "runASR_new_tree.txt", package="ASR")
#alternatively, specify the filename as a string
#new_tree <- "id_new_tree.txt"

save_tree(NULL, tree_file = new_tree) # if you want to use a specific tree file

save_tree(asrStructure, format = "png", name = "new_name")

}

