% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo_height_distrib}
\alias{logo_height_distrib}
\title{Calculate the height of letters in columns}
\usage{
logo_height_distrib(asrStructure, distribDF = NULL, alphabet = "AA")
}
\arguments{
\item{asrStructure}{the named list returned by \code{\link{runASR}} or \code{\link{loadASR}}. Set this to NULL
to specify other variables}

\item{distribDF}{the dataframe generated by \code{\link{read_distrib}}}
}
\value{
a dataframe with three columns representing the height of each AA at each column in the alignment. Formatted to
be passed to \code{\link{plot_logo}} or \code{\link{save_logo}}.\cr
heightDF$AA\cr
heightDF$Column \cr
heightDF$Height \cr
#Example\cr
heightDF[1, ] = "A", "5", "1.17" - An A in column 5 has a height of 1.17\cr
}
\description{
Takes the distribution data frame and calculates a logo height for each amino acid in each column
}
\examples{
data(asrStructure)

logo_height(asrStructure)

#if you want to use a specific set of data
#retrieve example file stored in the package
id_distribution <- system.file("extdata", "runASR_distribution.txt", package="ASR")
#alternatively, specify the filename as a string
#id_distribution <- "id_distribution.txt"

distribDF <- read_distrib(NULL, distrib_file = id_distribution)
logo_height(NULL, distribDF = distribDF) #if you want to specify your own file

}

